requirements.txt
Faker==35.2.0
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
six==1.17.0
typing_extensions==4.12.2
